chore(avatar, avatar-image): move to recipes
