+++ .storybook/recipes/AvatarImage/AvatarImage.test.ts
