+++ .storybook/recipes/AvatarImage/index.ts
